refactor(core): move hashing utilities out of lib
