+++ devtools/build_topology_projection.py
@@ -117,7 +117,6 @@ LIB_PREFIX_TO_SUBPACKAGE = {
 LIB_ROOT_PRIMITIVES = frozenset(
     {
         "__init__.py",
-        "hashing.py",
         "dates.py",
         "timestamps.py",
         "security.py",
