+++ polylogue/sources/parsers/base_support.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from collections.abc import Sequence
 
-from polylogue.lib.hashing import hash_text
+from polylogue.core.hashing import hash_text
 from polylogue.lib.roles import Role
 from polylogue.types import ContentBlockType
