+++ tests/benchmarks/test_pipeline.py
@@ -15,8 +15,8 @@ from sqlite3 import Connection
 import pytest
 
 from polylogue.archive.viewport.viewports import classify_tool
+from polylogue.core.hashing import hash_payload, hash_text
 from polylogue.core.json import JSONDocument
-from polylogue.lib.hashing import hash_payload, hash_text
 from polylogue.pipeline.prepare import PrepareCache
 from polylogue.pipeline.semantic_metadata import extract_tool_metadata
 from polylogue.storage.action_events.rebuild_runtime import (
